server/api/requirements.txt
# Web Framework
Django

# Database
psycopg2

# Security
argon2-cffi
django-filter
django-axes
django-guardian
django-oauth-toolkit
django-cors-middleware

# GraphQL
django-graphql-jwt
graphene-django
django-restql

# Google api
google-cloud-storage

# Data manipulation
Pillow
graphene-file-upload